refactor(#78): 로비 게임 시작 유스케이스 서비스 분리
